Fix status message position

+++ src/views/Edit.vue
@@ -1,17 +1,17 @@
 <template>
   <form class="text-left">
-    <h2>Edit</h2>
+    <h2>Edit Item</h2>
     <input type="text" v-model="todo.title" placeholder="Enter a Title" class="full-width text-larger" /><br/>
     <textarea v-model="todo.description" placeholder="Enter a Description" class="full-width"></textarea>
     <br/>
     <div class="text-right">
+      <code>{{ status }}</code>&nbsp;
       <button class="button" v-on:click="update">Update</button>&nbsp;
       <button class="button" v-on:click="remove">Delete</button>&nbsp;
-      <router-link to="/">Cancel</router-link>&nbsp;&nbsp;{{ status }}
+      <router-link to="/">Cancel</router-link>
     </div>
-    <br/><br/>
     Created: <code>{{ todo.created }}</code><br/>
-    Modified: <code>{{ todo.modified ? todo.modified : 'never' }}</code>
+    Modified: <code>{{ todo.modified }}</code>
   </form>
 </template>
 
@@ -41,6 +41,7 @@ export default {
       })
     },
     update () {
+      this.status = 'Saving...'
       const data = {
         '_id': this.todo._id,
         '_rev': this.todo._rev,
@@ -51,21 +52,28 @@ export default {
         'modified': new Date().toJSON()
       }
       store.update(data).then(() => {
-        this.todo = ''
-        this.status = 'Successfully Updated.'
+        store.read(data._id).then((doc) => {
+          this.todo = doc
+        })
+        this.status = 'Successfully Saved.'
       }).catch((err) => {
         console.log(err)
         this.status = err
       })
     },
     remove () {
-      store.delete(this.todo).then(() => {
-        this.todo = ''
-        this.status = 'Deleted.'
-      }).catch((err) => {
-        console.log(err)
-        this.status = err
-      })
+      this.status = ''
+      let isOk = confirm('Ok to delete this item?')
+      if(isOk) {
+        this.status = 'Deleting...'
+        store.delete(this.todo).then(() => {
+          this.todo = ''
+          this.status = 'Deleted.'
+        }).catch((err) => {
+          console.log(err)
+          this.status = err
+        })
+      }
     }
   }
 }
